fix problems with ints > 2^31
git-svn-id: svn+ssh://repo.cacti.net/var/data/svnroot/cacti/trunk/cactid_unix@2063 860744bd-22fc-0310-8c96-e9fe5004b5ca

+++ poller.c
@@ -224,7 +224,7 @@ void poll_host(int host_id) {
 					free(query3);
 
 					assert_fail = 1;
-				}else if ((!strcmp(reindex->op, ">")) && (atoi(reindex->assert_value) <= atoi(poll_result))) {
+				}else if ((!strcmp(reindex->op, ">")) && (atoll(reindex->assert_value) <= atoll(poll_result))) {
 					snprintf(logmessage, LOGSIZE, "ASSERT: '%s>%s' failed. Recaching host '%s', data query #%i\n", reindex->assert_value, poll_result, host->hostname, reindex->data_query_id);
 					cacti_log(logmessage);
 
@@ -234,7 +234,7 @@ void poll_host(int host_id) {
 					free(query3);
 
 					assert_fail = 1;
-				}else if ((!strcmp(reindex->op, "<")) && (atoi(reindex->assert_value) >= atoi(poll_result))) {
+				}else if ((!strcmp(reindex->op, "<")) && (atoll(reindex->assert_value) >= atoll(poll_result))) {
 					snprintf(logmessage, LOGSIZE, "ASSERT: '%s<%s' failed. Recaching host '%s', data query #%i\n", reindex->assert_value, poll_result, host->hostname, reindex->data_query_id);
 					cacti_log(logmessage);
